# orchestra_back/src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Orchestra API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "token",
      "in": "header",
      "name": "Authorization",
      "description": "The authorization is managed by a JWT Token provided on User login"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/${ENTITY_ROUTE}/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "description": "Log in for a User",
        "parameters": [
          {
            "name": "ENTITY_ROUTE",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "User email and password.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "UserId and JWT Token"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/${ENTITY_ROUTE}/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "description": "Log out for a User",
        "parameters": [
          {
            "name": "ENTITY_ROUTE",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success message"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/${ENTITY_ROUTE}/{projectId}/export": {
      "get": {
        "tags": [
          "Projects"
        ],
        "description": "Returns the whole information of a project into a single JSON object",
        "parameters": [
          {
            "name": "ENTITY_ROUTE",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "JSON object of the Project"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/${ENTITY_ROUTE}/{organizationId}/users/{usermail}": {
      "post": {
        "tags": [
          "Organizations"
        ],
        "description": "Add an User as member of the Organization by its email",
        "parameters": [
          {
            "name": "ENTITY_ROUTE",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "usermail",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success message"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/${ENTITY_ROUTE}/{organizationId}/users/{userId}": {
      "delete": {
        "tags": [
          "Organizations"
        ],
        "description": "Removes an User from an Organization",
        "parameters": [
          {
            "name": "ENTITY_ROUTE",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success message"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/${ENTITY_ROUTE}/{templateId}/elements/": {
      "get": {
        "tags": [
          "Custom Component Templates"
        ],
        "description": "Find the Elemenets which belong to a specific Custom Component Template",
        "parameters": [
          {
            "name": "ENTITY_ROUTE",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "templateId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Elements list"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    }
  },
  "definitions": {
    "Login": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "johngonzalez@javeriana.edu.co"
        },
        "password": {
          "type": "string",
          "example": "miContrasena123"
        }
      }
    },
...
  }
}